JavaScript: Cut dependency of `SourceNode::Range` on `flowsTo`.
I've replaced the heuristic for identifying strings in Angular source code with a simpler, syntactic heuristic that in practice should give us a superset of the old heuristic.

I've removed two additional subclasses of `SourceNode::Range` in `HTTP` and `LodashUnderscore`. Neither of these is semantically necessary for our own code. It could affect customer-written extensions, but is unlikely to be a problem in practice.